Increase max text message buffer size for websocket connections (#14)
What is the current behavior you want to change?

This change resolves an issue with WebSocket connections between the browser and the Java application server. The WS connection was closing immediately in Chrome and, when connecting over the nextwork, in Firefox. See
https://groups.google.com/d/topic/kurento/f2TRN2TIRwM/discussion
for more details.

What is the new behavior provided by this change?

Increasing the size of the WebSocket text buffer resolves the error.

How has this been tested?

I've tested the following examples in Chrome and Firefox:

    kurento-hello-world
    kurento-group-call
    kurento-one2many-call

This also resolved the same websocket issue for another developer in kurento-magic-mirror.